Backend/requirements.txt
streamlit==1.18.1
requests==2.31.0
python-dotenv==1.0.0
sounddevice==0.4.6
scipy==1.11.2
gTTS==2.3.2
SpeechRecognition==3.10.0
openai==0.27.8
langchain-openai==0.0.3
langchain==0.0.208
faiss-cpu==1.7.4
PyPDF2==3.0.1
python-docx==0.8.11
beautifulsoup4==4.12.2
urllib3==2.0.5
playsound==1.3.0